requirements.txt
annotated-types==0.6.0
anyio==4.2.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
distro==1.9.0
exceptiongroup==1.2.0
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-texttospeech==2.15.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
gTTS==2.5.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
lxml==4.9.4
moviepy==1.0.3
newsapi-python==0.2.7
numpy==1.26.2
openai==1.6.1
pandas==2.1.4
Pillow==10.1.0
proglog==0.1.10
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytrends==4.9.2
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
